+++ raster_loader/io/bigquery.py
@@ -6,8 +6,8 @@ import re
 
 from itertools import chain
 from raster_loader import __version__
-from raster_loader.errors import import_error_bigquery, IncompatibleRasterException
-from raster_loader.utils import ask_yes_no_question, batched
+from raster_loader.lib.errors import import_error_bigquery, IncompatibleRasterException
+from raster_loader.lib.utils import ask_yes_no_question, batched
 from raster_loader.io.common import (
     check_metadata_is_compatible,
     get_number_of_blocks,
